+++ src/actions/mixeffect/programPreview.ts
@@ -1,8 +1,8 @@
 import type { Atem } from 'atem-connection'
 import { getMixEffect } from 'atem-connection/dist/state/util.js'
-import { AtemMEPicker, AtemMESourcePicker } from '../../input.js'
+import { AtemMEPicker, AtemMESourcePicker, FadeDurationChoice } from '../../input.js'
 import type { ModelSpec } from '../../models/index'
-import { ActionId } from '../index.js'
+import { ActionId } from '../ActionId.js'
 import type { MyActionDefinitions } from '../types.js'
 import type { StateWrapper } from '../../state.js'
 import type { AtemTransitions } from '../../transitions.js'
