+++ src/actions/mixeffect/upstreamKeyerCommon.ts
@@ -7,7 +7,7 @@ import {
 	AtemUSKPicker,
 } from '../../input.js'
 import type { ModelSpec } from '../../models/index.js'
-import { ActionId } from '../index.js'
+import { ActionId } from '../ActionId.js'
 import type { MyActionDefinitions } from '../types.js'
 import { CHOICES_KEYTRANS, type TrueFalseToggle } from '../../choices.js'
 import { getUSK, type StateWrapper } from '../../state.js'
