+++ src/actions/settings.ts
@@ -1,6 +1,6 @@
 import { type Atem, type InputState } from 'atem-connection'
 import type { ModelSpec } from '../models/index.js'
-import { ActionId } from './index.js'
+import { ActionId } from './ActionId.js'
 import type { MyActionDefinitions } from './types.js'
 import { AtemAllSourcePicker } from '../input.js'
 import type { StateWrapper } from '../state.js'
