+++ src/actions/superSource.ts
@@ -1,6 +1,6 @@
 import { Enums, VideoState, type Atem } from 'atem-connection'
 import type { ModelSpec } from '../models/index.js'
-import { ActionId } from './index.js'
+import { ActionId } from './ActionId.js'
 import type { MyActionDefinitions } from './types.js'
 import {
 	AtemSuperSourceArtPropertiesPickers,
