+++ src/feedback/FeedbackId.ts
@@ -1,0 +1,54 @@
+export enum FeedbackId {
+	PreviewBG = 'preview_bg',
+	PreviewVariables = 'previewVariables',
+	PreviewBG2 = 'preview_bg_2',
+	PreviewBG3 = 'preview_bg_3',
+	PreviewBG4 = 'preview_bg_4',
+	ProgramBG = 'program_bg',
+	ProgramVariables = 'programVariables',
+	ProgramBG2 = 'program_bg_2',
+	ProgramBG3 = 'program_bg_3',
+	ProgramBG4 = 'program_bg_4',
+	AuxBG = 'aux_bg',
+	AuxVariables = 'auxVariables',
+	USKOnAir = 'usk_bg',
+	USKSource = 'usk_source',
+	USKSourceVariables = 'usk_source_variables',
+	USKKeyFrame = 'usk_keyframe',
+	DSKOnAir = 'dsk_bg',
+	DSKTie = 'dskTie',
+	DSKSource = 'dsk_source',
+	DSKSourceVariables = 'dsk_source_variables',
+	Macro = 'macro',
+	MacroLoop = 'macroloop',
+	MVSource = 'mv_source',
+	MVSourceVariables = 'mv_source_variables',
+	SSrcArtProperties = 'ssrc_art_properties',
+	SSrcArtSource = 'ssrc_art_source',
+	SSrcArtOption = 'ssrc_art_option',
+	SSrcBoxOnAir = 'ssrc_box_enable',
+	SSrcBoxSource = 'ssrc_box_source',
+	SSrcBoxSourceVariables = 'ssrc_box_source_variables',
+	SSrcBoxProperties = 'ssrc_box_properties',
+	TransitionStyle = 'transitionStyle',
+	TransitionSelection = 'transitionSelection',
+	TransitionRate = 'transitionRate',
+	InTransition = 'inTransition',
+	MediaPlayerSource = 'mediaPlayerSource',
+	FadeToBlackIsBlack = 'fadeToBlackIsBlack',
+	FadeToBlackRate = 'fadeToBlackRate',
+	ProgramTally = 'program_tally',
+	PreviewTally = 'preview_tally',
+	AdvancedTally = 'advanced_tally',
+	StreamStatus = 'streamStatus',
+	RecordStatus = 'recordStatus',
+	ClassicAudioGain = 'classicAudioGain',
+	ClassicAudioMixOption = 'classicAudioMixOption',
+	ClassicAudioMasterGain = 'classicAudioMasterGain',
+	FairlightAudioFaderGain = 'fairlightAudioFaderGain',
+	FairlightAudioInputGain = 'fairlightAudioInputGain',
+	FairlightAudioMixOption = 'fairlightAudioMixOption',
+	FairlightAudioMasterGain = 'fairlightAudioMasterGain',
+	FairlightAudioMonitorMasterMuted = 'fairlightAudioMonitorMasterMuted',
+	FairlightAudioMonitorFaderGain = 'fairlightAudioMonitorFaderGain',
+}
