+++ src/feedback/dsk.ts
@@ -1,6 +1,6 @@
 import type { ModelSpec } from '../models/index.js'
 import type { MyFeedbackDefinitions } from './types.js'
-import { FeedbackId } from './index.js'
+import { FeedbackId } from './FeedbackId.js'
 import { combineRgb } from '@companion-module/base'
 import { AtemDSKPicker, AtemKeyFillSourcePicker } from '../input.js'
 import { getDSK, type StateWrapper } from '../state.js'
