+++ src/feedback/macro.ts
@@ -1,6 +1,6 @@
 import type { ModelSpec } from '../models/index.js'
 import type { MyFeedbackDefinitions } from './types.js'
-import { FeedbackId } from './index.js'
+import { FeedbackId } from './FeedbackId.js'
 import { combineRgb, type CompanionInputFieldDropdown } from '@companion-module/base'
 import { GetMacroChoices } from '../choices.js'
 import { assertUnreachable } from '../util.js'
