+++ src/feedback/mediaPlayer.ts
@@ -2,7 +2,7 @@ import { Enums } from 'atem-connection'
 import { AtemMediaPlayerPicker, AtemMediaPlayerSourcePicker } from '../input.js'
 import type { ModelSpec } from '../models/index.js'
 import type { MyFeedbackDefinitions } from './types.js'
-import { FeedbackId } from './index.js'
+import { FeedbackId } from './FeedbackId.js'
 import { combineRgb } from '@companion-module/base'
 import { MEDIA_PLAYER_SOURCE_CLIP_OFFSET } from '../util.js'
 import type { StateWrapper } from '../state.js'
