+++ src/feedback/mixeffect/tally.ts
@@ -1,7 +1,7 @@
 import { AtemMESourcePicker } from '../../input.js'
 import type { ModelSpec } from '../../models/index.js'
 import type { MyFeedbackDefinitions } from '../types.js'
-import { FeedbackId } from '../index.js'
+import { FeedbackId } from '../FeedbackId.js'
 import { combineRgb } from '@companion-module/base'
 import type { StateWrapper } from '../../state.js'
 import { calculateTallyForInputId } from '../../util.js'
