+++ src/feedback/mixeffect/usk.ts
@@ -2,7 +2,7 @@ import type { Enums } from 'atem-connection'
 import { AtemKeyFillSourcePicker, AtemMEPicker, AtemUSKPicker } from '../../input.js'
 import type { ModelSpec } from '../../models/index.js'
 import type { MyFeedbackDefinitions } from '../types.js'
-import { FeedbackId } from '../index.js'
+import { FeedbackId } from '../FeedbackId.js'
 import { combineRgb, type CompanionInputFieldDropdown } from '@companion-module/base'
 import { getUSK, type StateWrapper } from '../../state.js'
 import { CHOICES_CURRENTKEYFRAMES } from '../../choices.js'
