+++ src/index.ts
@@ -1,7 +1,8 @@
 import AtemPkg, { type Atem as IAtem, type AtemState, Commands } from 'atem-connection'
 import { GetActionsList } from './actions/index.js'
 import { type AtemConfig, GetConfigFields } from './config.js'
-import { FeedbackId, GetFeedbacksList } from './feedback/index.js'
+import { GetFeedbacksList } from './feedback/index.js'
+import { FeedbackId } from './feedback/FeedbackId.js'
 import { GetAutoDetectModel, GetModelSpec, GetParsedModelSpec, type ModelSpec } from './models/index.js'
 import { GetPresetsList } from './presets.js'
 import type { StateWrapper } from './state.js'
